+++ templates/contest_view_add_events.php
@@ -1,6 +1,6 @@
 <?php
 
-if (current_user_can(VESPA_Roles::versenyek_kezelese_kiiras_modositas_torles)) :
+if (vespa_user_can_edit_contest($id)) :
     $sports       = $wpdb->get_results("SELECT * FROM vespa_sports WHERE is_deleted=0 ORDER BY sport_name ASC");
     $events       = $wpdb->get_results("SELECT * FROM vespa_sport_events WHERE is_deleted=0 ORDER BY sport_event_name ASC");
     $disabilities = $wpdb->get_results("SELECT * FROM vespa_disability_groups ORDER BY disability_group_name ASC");